postgres_to_es/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==4.2.5
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
elastic-transport==8.11.0
elasticsearch==8.11.1
exceptiongroup==1.2.0
Faker==20.1.0
flake8==6.1.0
gunicorn==20.1.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
urllib3==2.1.0
uWSGI==2.0.21
